requirements.txt
beautifulsoup4==4.12.2
langchain==0.0.220
openai==0.27.4
python-dotenv==1.0.0
python_docx==0.8.11
Requests==2.31.0
streamlit==1.24.0